QGIS algorithm - Export to PostgreSQL (new connection)

QGIS Algorithm provided by GDAL Export to PostgreSQL (new connection) (gdal:importvectorintopostgisdatabasenewconnection). Exports a vector layer to a new PostgreSQL database connection

gdal_importvectorintopostgisdatabasenewconnection(
  INPUT = qgisprocess:::qgis_default_value(),
  SHAPE_ENCODING = qgisprocess:::qgis_default_value(),
  GTYPE = qgisprocess:::qgis_default_value(),
  A_SRS = qgisprocess:::qgis_default_value(),
  T_SRS = qgisprocess:::qgis_default_value(),
  S_SRS = qgisprocess:::qgis_default_value(),
  HOST = qgisprocess:::qgis_default_value(),
  PORT = qgisprocess:::qgis_default_value(),
  USER = qgisprocess:::qgis_default_value(),
  DBNAME = qgisprocess:::qgis_default_value(),
  PASSWORD = qgisprocess:::qgis_default_value(),
  SCHEMA = qgisprocess:::qgis_default_value(),
  TABLE = qgisprocess:::qgis_default_value(),
  PK = qgisprocess:::qgis_default_value(),
  PRIMARY_KEY = qgisprocess:::qgis_default_value(),
  GEOCOLUMN = qgisprocess:::qgis_default_value(),
  DIM = qgisprocess:::qgis_default_value(),
  SIMPLIFY = qgisprocess:::qgis_default_value(),
  SEGMENTIZE = qgisprocess:::qgis_default_value(),
  SPAT = qgisprocess:::qgis_default_value(),
  CLIP = qgisprocess:::qgis_default_value(),
  FIELDS = qgisprocess:::qgis_default_value(),
  WHERE = qgisprocess:::qgis_default_value(),
  GT = qgisprocess:::qgis_default_value(),
  OVERWRITE = qgisprocess:::qgis_default_value(),
  APPEND = qgisprocess:::qgis_default_value(),
  ADDFIELDS = qgisprocess:::qgis_default_value(),
  LAUNDER = qgisprocess:::qgis_default_value(),
  INDEX = qgisprocess:::qgis_default_value(),
  SKIPFAILURES = qgisprocess:::qgis_default_value(),
  MAKEVALID = qgisprocess:::qgis_default_value(),
  PROMOTETOMULTI = qgisprocess:::qgis_default_value(),
  PRECISION = qgisprocess:::qgis_default_value(),
  OPTIONS = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)

Arguments

INPUT

source - Input layer. Path to a vector layer.

SHAPE_ENCODING

string - Shape encoding. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

GTYPE

enum of ("", "NONE", "GEOMETRY", "POINT", "LINESTRING", "POLYGON", "GEOMETRYCOLLECTION", "MULTIPOINT", "MULTIPOLYGON", "MULTILINESTRING", "CIRCULARSTRING", "COMPOUNDCURVE", "CURVEPOLYGON", "MULTICURVE", "MULTISURFACE", "CONVERT_TO_LINEAR", "CONVERT_TO_CURVE") - Output geometry type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.

A_SRS

crs - Assign an output CRS. CRS as an auth ID (e.g. 'EPSG:3111'). CRS as a PROJ4 string (e.g. 'PROJ4:…'). CRS as a WKT string (e.g. 'WKT:…'). Path to a layer. The CRS of the layer is used..

T_SRS

crs - Reproject to this CRS on output . CRS as an auth ID (e.g. 'EPSG:3111'). CRS as a PROJ4 string (e.g. 'PROJ4:…'). CRS as a WKT string (e.g. 'WKT:…'). Path to a layer. The CRS of the layer is used..

S_SRS

crs - Override source CRS. CRS as an auth ID (e.g. 'EPSG:3111'). CRS as a PROJ4 string (e.g. 'PROJ4:…'). CRS as a WKT string (e.g. 'WKT:…'). Path to a layer. The CRS of the layer is used..

HOST

string - Host. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

PORT

string - Port. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

USER

string - Username. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

DBNAME

string - Database name. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

PASSWORD

string - Password. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

SCHEMA

string - Schema name. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

TABLE

string - Table name, leave blank to use input name. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

PK

string - Primary key (new field). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

PRIMARY_KEY

field - Primary key (existing field, used if the above option is left empty). The name of an existing field. ; delimited list of existing field names.

GEOCOLUMN

string - Geometry column name. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

DIM

enum of ("2", "3", "4") - Vector dimensions. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.

SIMPLIFY

string - Distance tolerance for simplification. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

SEGMENTIZE

string - Maximum distance between 2 nodes (densification). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

SPAT

extent - Select features by extent (defined in input layer CRS). A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..

CLIP

boolean - Clip the input layer using the above (rectangle) extent. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

FIELDS

field - Fields to include (leave empty to use all fields). The name of an existing field. ; delimited list of existing field names.

WHERE

string - Select features using a SQL "WHERE" statement (Ex: column='value'). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

GT

string - Group N features per transaction (Default: 20000). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

OVERWRITE

boolean - Overwrite existing table. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

APPEND

boolean - Append to existing table. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

ADDFIELDS

boolean - Append and add new fields to existing table. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

LAUNDER

boolean - Do not launder columns/table names. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

INDEX

boolean - Do not create spatial index. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

SKIPFAILURES

boolean - Continue after a failure, skipping the failed feature. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

MAKEVALID

boolean - Validate geometries based on Simple Features specification. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

PROMOTETOMULTI

boolean - Promote to Multipart. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

PRECISION

boolean - Keep width and precision of input attributes. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

OPTIONS

string - Additional creation options. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

...

further parameters passed to qgisprocess::qgis_run_algorithm()

.complete_output

logical specifying if complete out of qgisprocess::qgis_run_algorithm() should be used (TRUE) or first output (most likely the main) should read (FALSE). Default value is TRUE.

.quiet

logical specifying if parameter .quiet for qgisprocess::qgis_run_algorithm() Default value is TRUE.

.messages

logical specifying if messages from qgisprocess::qgis_run_algorithm() should be printed (TRUE) or not (FALSE). Default value is FALSE.

Details

Outputs description

No output description provided.