QGIS algorithm Proximity (raster distance)

QGIS Algorithm provided by GDAL Proximity (raster distance) (gdal:proximity)

gdal_proximity(
  INPUT = qgisprocess::qgis_default_value(),
  BAND = qgisprocess::qgis_default_value(),
  VALUES = qgisprocess::qgis_default_value(),
  UNITS = qgisprocess::qgis_default_value(),
  MAX_DISTANCE = qgisprocess::qgis_default_value(),
  REPLACE = qgisprocess::qgis_default_value(),
  NODATA = qgisprocess::qgis_default_value(),
  OPTIONS = qgisprocess::qgis_default_value(),
  EXTRA = qgisprocess::qgis_default_value(),
  DATA_TYPE = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)

Arguments

INPUT
raster - Input layer. Path to a raster layer.

BAND
band - Band number. Integer value representing an existing raster band number.

VALUES
string - A list of pixel values in the source image to be considered target pixels. String value.

UNITS
enum of ("Georeferenced coordinates", "Pixel coordinates") - Distance units. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.

MAX_DISTANCE
number - The maximum distance to be generated. A numeric value.

REPLACE
number - Value to be applied to all pixels that are within the -maxdist of target pixels. A numeric value.

NODATA
number - Nodata value to use for the destination proximity raster. A numeric value.

OPTIONS
string - Additional creation options. String value.

EXTRA
string - Additional command-line parameters. String value.

DATA_TYPE
enum of ("Byte", "Int16", "UInt16", "UInt32", "Int32", "Float32", "Float64", "CInt16", "CInt32", "CFloat32", "CFloat64") - Output data type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.

OUTPUT
rasterDestination - Proximity map. Path for new raster layer.

...
further parameters passed to qgisprocess::qgis_run_algorithm()

.complete_output
logical specifing if complete out of qgisprocess::qgis_run_algorithm() should be used (TRUE) or first output (most likely the main) should read (FALSE). Default value is TRUE.

Details

Outputs description

OUTPUT - outputRaster - Proximity map