QGIS algorithm - Warp (reproject)

QGIS Algorithm provided by GDAL Warp (reproject) (gdal:warpreproject). —————- Arguments —————- INPUT: Input layer Argument type: raster Acceptable values: - Path to a raster layer SOURCE_CRS: Source CRS (optional) Argument type: crs Acceptable values: - CRS as an auth ID (e.g. 'EPSG:3111') - CRS as a PROJ4 string (e.g. 'PROJ4:…') - CRS as a WKT string (e.g. 'WKT:…') - Path to a layer. The CRS of the layer is used. TARGET_CRS: Target CRS (optional) Argument type: crs Acceptable values: - CRS as an auth ID (e.g. 'EPSG:3111') - CRS as a PROJ4 string (e.g. 'PROJ4:…') - CRS as a WKT string (e.g. 'WKT:…') - Path to a layer. The CRS of the layer is used. RESAMPLING: Resampling method to use Default value: 0 Argument type: enum Available values: - 0: Nearest Neighbour - 1: Bilinear (2x2 Kernel) - 2: Cubic (4x4 Kernel) - 3: Cubic B-Spline (4x4 Kernel) - 4: Lanczos (6x6 Kernel) - 5: Average - 6: Mode - 7: Maximum - 8: Minimum - 9: Median - 10: First Quartile (Q1) - 11: Third Quartile (Q3) Acceptable values: - Number of selected option, e.g. '1' - Comma separated list of options, e.g. '1,3' NODATA: Nodata value for output bands (optional) Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression TARGET_RESOLUTION: Output file resolution in target georeferenced units (optional) Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression OPTIONS: Additional creation options (optional) Default value: Argument type: string Acceptable values: - String value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DATA_TYPE: Output data type Default value: 0 Argument type: enum Available values: - 0: Use Input Layer Data Type - 1: Byte - 2: Int16 - 3: UInt16 - 4: UInt32 - 5: Int32 - 6: Float32 - 7: Float64 - 8: CInt16 - 9: CInt32 - 10: CFloat32 - 11: CFloat64 - 12: Int8 Acceptable values: - Number of selected option, e.g. '1' - Comma separated list of options, e.g. '1,3' TARGET_EXTENT: Georeferenced extents of output file to be created (optional) Argument type: extent Acceptable values: - A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105' - Path to a layer. The extent of the layer is used. TARGET_EXTENT_CRS: CRS of the target raster extent (optional) Argument type: crs Acceptable values: - CRS as an auth ID (e.g. 'EPSG:3111') - CRS as a PROJ4 string (e.g. 'PROJ4:…') - CRS as a WKT string (e.g. 'WKT:…') - Path to a layer. The CRS of the layer is used. MULTITHREADING: Use multithreaded warping implementation Default value: false Argument type: boolean Acceptable values: - 1 for true/yes - 0 for false/no - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression EXTRA: Additional command-line parameters (optional) Argument type: string Acceptable values: - String value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression OUTPUT: Reprojected Argument type: rasterDestination Acceptable values: - Path for new raster layer

gdal_warpreproject(
  INPUT = qgisprocess:::qgis_default_value(),
  SOURCE_CRS = qgisprocess:::qgis_default_value(),
  TARGET_CRS = qgisprocess:::qgis_default_value(),
  RESAMPLING = qgisprocess:::qgis_default_value(),
  NODATA = qgisprocess:::qgis_default_value(),
  TARGET_RESOLUTION = qgisprocess:::qgis_default_value(),
  OPTIONS = qgisprocess:::qgis_default_value(),
  DATA_TYPE = qgisprocess:::qgis_default_value(),
  TARGET_EXTENT = qgisprocess:::qgis_default_value(),
  TARGET_EXTENT_CRS = qgisprocess:::qgis_default_value(),
  MULTITHREADING = qgisprocess:::qgis_default_value(),
  EXTRA = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)

Arguments

INPUT

raster - Input layer. Path to a raster layer.

SOURCE_CRS

crs - Source CRS. CRS as an auth ID (e.g. 'EPSG:3111'). CRS as a PROJ4 string (e.g. 'PROJ4:…'). CRS as a WKT string (e.g. 'WKT:…'). Path to a layer. The CRS of the layer is used..

TARGET_CRS

crs - Target CRS. CRS as an auth ID (e.g. 'EPSG:3111'). CRS as a PROJ4 string (e.g. 'PROJ4:…'). CRS as a WKT string (e.g. 'WKT:…'). Path to a layer. The CRS of the layer is used..

RESAMPLING

enum of ("Nearest Neighbour", "Bilinear (2x2 Kernel)", "Cubic (4x4 Kernel)", "Cubic B-Spline (4x4 Kernel)", "Lanczos (6x6 Kernel)", "Average", "Mode", "Maximum", "Minimum", "Median", "First Quartile (Q1)", "Third Quartile (Q3)") - Resampling method to use. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.

NODATA

number - Nodata value for output bands. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

TARGET_RESOLUTION

number - Output file resolution in target georeferenced units. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

OPTIONS

string - Additional creation options. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

DATA_TYPE

enum of ("Use Input Layer Data Type", "Byte", "Int16", "UInt16", "UInt32", "Int32", "Float32", "Float64", "CInt16", "CInt32", "CFloat32", "CFloat64", "Int8") - Output data type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.

TARGET_EXTENT

extent - Georeferenced extents of output file to be created. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..

TARGET_EXTENT_CRS

crs - CRS of the target raster extent. CRS as an auth ID (e.g. 'EPSG:3111'). CRS as a PROJ4 string (e.g. 'PROJ4:…'). CRS as a WKT string (e.g. 'WKT:…'). Path to a layer. The CRS of the layer is used..

MULTITHREADING

boolean - Use multithreaded warping implementation. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

EXTRA

string - Additional command-line parameters. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

OUTPUT

rasterDestination - Reprojected. Path for new raster layer.

...

further parameters passed to qgisprocess::qgis_run_algorithm()

.complete_output

logical specifying if complete out of qgisprocess::qgis_run_algorithm() should be used (TRUE) or first output (most likely the main) should read (FALSE). Default value is TRUE.

.quiet

logical specifying if parameter .quiet for qgisprocess::qgis_run_algorithm() Default value is TRUE.

.messages

logical specifying if messages from qgisprocess::qgis_run_algorithm() should be printed (TRUE) or not (FALSE). Default value is FALSE.

Details

Outputs description

OUTPUT - outputRaster - Reprojected