QGIS algorithm i.evapo.pm

QGIS Algorithm provided by GRASS i.evapo.pm (grass7:i.evapo.pm)

grass7_i_evapo_pm(
  elevation = qgisprocess::qgis_default_value(),
  temperature = qgisprocess::qgis_default_value(),
  relativehumidity = qgisprocess::qgis_default_value(),
  windspeed = qgisprocess::qgis_default_value(),
  netradiation = qgisprocess::qgis_default_value(),
  cropheight = qgisprocess::qgis_default_value(),
  .z = qgisprocess::qgis_default_value(),
  .n = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)

Arguments

elevation

raster - Name of input elevation raster map (m a.s.l.). Path to a raster layer.

temperature

raster - Name of input temperature raster map (C). Path to a raster layer.

relativehumidity

raster - Name of input relative humidity raster map (%). Path to a raster layer.

windspeed

raster - Name of input wind speed raster map (m/s). Path to a raster layer.

netradiation

raster - Name of input net solar radiation raster map (MJ/m2/h). Path to a raster layer.

cropheight

raster - Name of input crop height raster map (m). Path to a raster layer.

.z

boolean - Set negative ETa to zero. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -z.

.n

boolean - Use Night-time. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -n.

output

rasterDestination - Evapotranspiration. Path for new raster layer.

GRASS_REGION_PARAMETER

extent - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..

GRASS_REGION_CELLSIZE_PARAMETER

number - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.

GRASS_RASTER_FORMAT_OPT

string - Output Rasters format options (createopt). String value.

GRASS_RASTER_FORMAT_META

string - Output Rasters format metadata options (metaopt). String value.

...

further parameters passed to qgisprocess::qgis_run_algorithm()

.complete_output

logical specifing if complete out of qgisprocess::qgis_run_algorithm() should be used (TRUE) or first output (most likely the main) should read (FALSE). Default value is TRUE.

Details

Outputs description

output - outputRaster - Evapotranspiration