QGIS algorithm r.recode

QGIS Algorithm provided by GRASS r.recode (grass7:r.recode)

grass7_r_recode(
  input = qgisprocess::qgis_default_value(),
  rules = qgisprocess::qgis_default_value(),
  .d = qgisprocess::qgis_default_value(),
  .a = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)

Arguments

input
raster - Input layer. Path to a raster layer.

rules
file - File containing recode rules. Path to a file.

.d
boolean - Force output to 'double' raster map type (DCELL). 1 for true/yes. 0 for false/no. Original algorithm parameter name: name.

.a
boolean - Align the current region to the input raster map. 1 for true/yes. 0 for false/no. Original algorithm parameter name: name.

output
rasterDestination - Recoded. Path for new raster layer.

GRASS_REGION_PARAMETER
extent - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..

GRASS_REGION_CELLSIZE_PARAMETER
number - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.

GRASS_RASTER_FORMAT_OPT
string - Output Rasters format options (createopt). String value.

GRASS_RASTER_FORMAT_META
string - Output Rasters format metadata options (metaopt). String value.

...
further parameters passed to qgisprocess::qgis_run_algorithm()

.complete_output
logical specifing if complete out of qgisprocess::qgis_run_algorithm() should be used (TRUE) or first output (most likely the main) should read (FALSE). Default value is TRUE.

Details

Outputs description

output - outputRaster - Recoded