QGIS algorithm v.generalize

QGIS Algorithm provided by GRASS v.generalize (grass7:v.generalize)

grass7_v_generalize(
  input = qgisprocess::qgis_default_value(),
  type = qgisprocess::qgis_default_value(),
  cats = qgisprocess::qgis_default_value(),
  where = qgisprocess::qgis_default_value(),
  method = qgisprocess::qgis_default_value(),
  threshold = qgisprocess::qgis_default_value(),
  look_ahead = qgisprocess::qgis_default_value(),
  reduction = qgisprocess::qgis_default_value(),
  slide = qgisprocess::qgis_default_value(),
  angle_thresh = qgisprocess::qgis_default_value(),
  degree_thresh = qgisprocess::qgis_default_value(),
  closeness_thresh = qgisprocess::qgis_default_value(),
  betweeness_thresh = qgisprocess::qgis_default_value(),
  alpha = qgisprocess::qgis_default_value(),
  beta = qgisprocess::qgis_default_value(),
  iterations = qgisprocess::qgis_default_value(),
  .t = qgisprocess::qgis_default_value(),
  .l = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  error = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_SNAP_TOLERANCE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_MIN_AREA_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_OUTPUT_TYPE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_DSCO = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_LCO = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_EXPORT_NOCAT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)

Arguments

input
source - Input layer. Path to a vector layer.

type
enum of ("line", "boundary", "area") - Input feature type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.

cats
string - Category values. String value.

where
string - WHERE conditions of SQL statement without 'where' keyword. String value.

method
enum of ("douglas", "douglas_reduction", "lang", "reduction", "reumann", "boyle", "sliding_averaging", "distance_weighting", "chaiken", "hermite", "snakes", "network", "displacement") - Generalization algorithm. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.

threshold
number - Maximal tolerance value. A numeric value.

look_ahead
number - Look-ahead parameter. A numeric value.

reduction
number - Percentage of the points in the output of 'douglas_reduction' algorithm. A numeric value.

slide
number - Slide of computed point toward the original point. A numeric value.

angle_thresh
number - Minimum angle between two consecutive segments in Hermite method. A numeric value.

degree_thresh
number - Degree threshold in network generalization. A numeric value.

closeness_thresh
number - Closeness threshold in network generalization. A numeric value.

betweeness_thresh
number - Betweenness threshold in network generalization. A numeric value.

alpha
number - Snakes alpha parameter. A numeric value.

beta
number - Snakes beta parameter. A numeric value.

iterations
number - Number of iterations. A numeric value.

.t
boolean - Do not copy attributes. 1 for true/yes. 0 for false/no. Original algorithm parameter name: name.

.l
boolean - Disable loop support. 1 for true/yes. 0 for false/no. Original algorithm parameter name: name.

output
vectorDestination - Generalized. Path for new vector layer.

error
vectorDestination - Errors. Path for new vector layer.

GRASS_REGION_PARAMETER
extent - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..

GRASS_SNAP_TOLERANCE_PARAMETER
number - v.in.ogr snap tolerance (-1 = no snap). A numeric value.

GRASS_MIN_AREA_PARAMETER
number - v.in.ogr min area. A numeric value.

GRASS_OUTPUT_TYPE_PARAMETER
enum of ("auto", "point", "line", "area") - v.out.ogr output type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.

GRASS_VECTOR_DSCO
string - v.out.ogr output data source options (dsco). String value.

GRASS_VECTOR_LCO
string - v.out.ogr output layer options (lco). String value.

GRASS_VECTOR_EXPORT_NOCAT
boolean - Also export features without category (not labeled). Otherwise only features with category are exported. 1 for true/yes. 0 for false/no.

...
further parameters passed to qgisprocess::qgis_run_algorithm()

.complete_output
logical specifing if complete out of qgisprocess::qgis_run_algorithm() should be used (TRUE) or first output (most likely the main) should read (FALSE). Default value is TRUE.

Details

Outputs description

output - outputVector - Generalized

error - outputVector - Errors