QGIS algorithm - r.out.mpeg

QGIS Algorithm provided by GRASS r.out.mpeg (grass:r.out.mpeg). Converts raster map series to MPEG movie

grass_r_out_mpeg(
  view1 = qgisprocess:::qgis_default_value(),
  view2 = qgisprocess:::qgis_default_value(),
  view3 = qgisprocess:::qgis_default_value(),
  view4 = qgisprocess:::qgis_default_value(),
  quality = qgisprocess:::qgis_default_value(),
  output = qgisprocess:::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)

Arguments

view1

multilayer - Name of input raster map(s) for view no.1. .

view2

multilayer - Name of input raster map(s) for view no.2. .

view3

multilayer - Name of input raster map(s) for view no.3. .

view4

multilayer - Name of input raster map(s) for view no.4. .

quality

number - Quality factor (1 = highest quality, lowest compression). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

output

fileDestination - MPEG file. Path for new file.

GRASS_REGION_PARAMETER

extent - GRASS GIS region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..

GRASS_REGION_CELLSIZE_PARAMETER

number - GRASS GIS region cellsize (leave 0 for default). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

...

further parameters passed to qgisprocess::qgis_run_algorithm()

.complete_output

logical specifying if complete out of qgisprocess::qgis_run_algorithm() should be used (TRUE) or first output (most likely the main) should read (FALSE). Default value is TRUE.

.quiet

logical specifying if parameter .quiet for qgisprocess::qgis_run_algorithm() Default value is TRUE.

.messages

logical specifying if messages from qgisprocess::qgis_run_algorithm() should be printed (TRUE) or not (FALSE). Default value is FALSE.

Details

Outputs description

output - outputFile - MPEG file