QGIS algorithm - r.slope.aspect

QGIS Algorithm provided by GRASS r.slope.aspect (grass:r.slope.aspect). Generates raster layers of slope, aspect, curvatures and partial derivatives from a elevation raster layer.

grass_r_slope_aspect(
  elevation = qgisprocess:::qgis_default_value(),
  format = qgisprocess:::qgis_default_value(),
  precision = qgisprocess:::qgis_default_value(),
  .a = qgisprocess:::qgis_default_value(),
  .e = qgisprocess:::qgis_default_value(),
  .n = qgisprocess:::qgis_default_value(),
  zscale = qgisprocess:::qgis_default_value(),
  min_slope = qgisprocess:::qgis_default_value(),
  slope = qgisprocess:::qgis_default_value(),
  aspect = qgisprocess:::qgis_default_value(),
  pcurvature = qgisprocess:::qgis_default_value(),
  tcurvature = qgisprocess:::qgis_default_value(),
  dx = qgisprocess:::qgis_default_value(),
  dy = qgisprocess:::qgis_default_value(),
  dxx = qgisprocess:::qgis_default_value(),
  dyy = qgisprocess:::qgis_default_value(),
  dxy = qgisprocess:::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess:::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)

Arguments

elevation

raster - Elevation. Path to a raster layer.

format

enum of ("degrees", "percent") - Format for reporting the slope. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.

precision

enum of ("FCELL", "CELL", "DCELL") - Type of output aspect and slope layer. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.

.a

boolean - Do not align the current region to the elevation layer. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression. Original algorithm parameter name: -a.

.e

boolean - Compute output at edges and near NULL values. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression. Original algorithm parameter name: -e.

.n

boolean - Create aspect as degrees clockwise from North (azimuth), with flat = -9999. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression. Original algorithm parameter name: -n.

zscale

number - Multiplicative factor to convert elevation units to meters. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

min_slope

number - Minimum slope val. (in percent) for which aspect is computed. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

slope

rasterDestination - Slope. Path for new raster layer.

aspect

rasterDestination - Aspect. Path for new raster layer.

pcurvature

rasterDestination - Profile curvature. Path for new raster layer.

tcurvature

rasterDestination - Tangential curvature. Path for new raster layer.

dx

rasterDestination - First order partial derivative dx (E-W slope). Path for new raster layer.

dy

rasterDestination - First order partial derivative dy (N-S slope). Path for new raster layer.

dxx

rasterDestination - Second order partial derivative dxx. Path for new raster layer.

dyy

rasterDestination - Second order partial derivative dyy. Path for new raster layer.

dxy

rasterDestination - Second order partial derivative dxy. Path for new raster layer.

GRASS_REGION_PARAMETER

extent - GRASS GIS region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..

GRASS_REGION_CELLSIZE_PARAMETER

number - GRASS GIS region cellsize (leave 0 for default). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

GRASS_RASTER_FORMAT_OPT

string - Output Rasters format options (createopt). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

GRASS_RASTER_FORMAT_META

string - Output Rasters format metadata options (metaopt). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

...

further parameters passed to qgisprocess::qgis_run_algorithm()

.complete_output

logical specifying if complete out of qgisprocess::qgis_run_algorithm() should be used (TRUE) or first output (most likely the main) should read (FALSE). Default value is TRUE.

.quiet

logical specifying if parameter .quiet for qgisprocess::qgis_run_algorithm() Default value is TRUE.

.messages

logical specifying if messages from qgisprocess::qgis_run_algorithm() should be printed (TRUE) or not (FALSE). Default value is FALSE.

Details

Outputs description

aspect - outputRaster - Aspect

dx - outputRaster - First order partial derivative dx (E-W slope)

dxx - outputRaster - Second order partial derivative dxx

dxy - outputRaster - Second order partial derivative dxy

dy - outputRaster - First order partial derivative dy (N-S slope)

dyy - outputRaster - Second order partial derivative dyy

pcurvature - outputRaster - Profile curvature

slope - outputRaster - Slope

tcurvature - outputRaster - Tangential curvature