QGIS algorithm - r.solute.transport

QGIS Algorithm provided by GRASS r.solute.transport (grass:r.solute.transport). Numerical calculation program for transient, confined and unconfined solute transport in two dimensions

grass_r_solute_transport(
  c = qgisprocess:::qgis_default_value(),
  phead = qgisprocess:::qgis_default_value(),
  hc_x = qgisprocess:::qgis_default_value(),
  hc_y = qgisprocess:::qgis_default_value(),
  status = qgisprocess:::qgis_default_value(),
  diff_x = qgisprocess:::qgis_default_value(),
  diff_y = qgisprocess:::qgis_default_value(),
  q = qgisprocess:::qgis_default_value(),
  cin = qgisprocess:::qgis_default_value(),
  cs = qgisprocess:::qgis_default_value(),
  rd = qgisprocess:::qgis_default_value(),
  nf = qgisprocess:::qgis_default_value(),
  top = qgisprocess:::qgis_default_value(),
  bottom = qgisprocess:::qgis_default_value(),
  dtime = qgisprocess:::qgis_default_value(),
  maxit = qgisprocess:::qgis_default_value(),
  error = qgisprocess:::qgis_default_value(),
  solver = qgisprocess:::qgis_default_value(),
  relax = qgisprocess:::qgis_default_value(),
  al = qgisprocess:::qgis_default_value(),
  at = qgisprocess:::qgis_default_value(),
  loops = qgisprocess:::qgis_default_value(),
  stab = qgisprocess:::qgis_default_value(),
  .c = qgisprocess:::qgis_default_value(),
  .f = qgisprocess:::qgis_default_value(),
  output = qgisprocess:::qgis_default_value(),
  vx = qgisprocess:::qgis_default_value(),
  vy = qgisprocess:::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess:::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)

Arguments

c

raster - The initial concentration in (kg/m^3). Path to a raster layer.

phead

raster - The piezometric head in (m). Path to a raster layer.

hc_x

raster - The x-part of the hydraulic conductivity tensor in (m/s). Path to a raster layer.

hc_y

raster - The y-part of the hydraulic conductivity tensor in (m/s). Path to a raster layer.

status

raster - The status for each cell, = 0 - inactive cell, 1 - active cell, 2 - dirichlet- and 3 - transfer boundary condition. Path to a raster layer.

diff_x

raster - The x-part of the diffusion tensor in (m^2/s). Path to a raster layer.

diff_y

raster - The y-part of the diffusion tensor in (m^2/s). Path to a raster layer.

q

raster - Groundwater sources and sinks in (m^3/s). Path to a raster layer.

cin

raster - Concentration sources and sinks bounded to a water source or sink in (kg/s). Path to a raster layer.

cs

raster - Concentration of inner sources and inner sinks in (kg/s) (i.e. a chemical reaction). Path to a raster layer.

rd

raster - Retardation factor (-). Path to a raster layer.

nf

raster - Effective porosity (-). Path to a raster layer.

top

raster - Top surface of the aquifer in (m). Path to a raster layer.

bottom

raster - Bottom surface of the aquifer in (m). Path to a raster layer.

dtime

number - Calculation time (in seconds). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

maxit

number - Maximum number of iteration used to solve the linear equation system. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

error

number - Error break criteria for iterative solver. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

solver

enum of ("gauss", "lu", "jacobi", "sor", "bicgstab") - The type of solver which should solve the linear equation system. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.

relax

number - The relaxation parameter used by the jacobi and sor solver for speedup or stabilizing. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

al

number - The longitudinal dispersivity length. (m). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

at

number - The transversal dispersivity length. (m). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

loops

number - Use this number of time loops if the CFL flag is off. The timestep will become dt/loops.. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

stab

enum of ("full", "exp") - Set the flow stabilizing scheme (full or exponential upwinding).. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.

.c

boolean - Use the Courant-Friedrichs-Lewy criteria for time step calculation. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression. Original algorithm parameter name: -c.

.f

boolean - Use a full filled quadratic linear equation system, default is a sparse linear equation system.. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression. Original algorithm parameter name: -f.

output

rasterDestination - Solute Transport. Path for new raster layer.

vx

rasterDestination - Calculate and store the groundwater filter velocity vector part in x direction (m/s). Path for new raster layer.

vy

rasterDestination - Calculate and store the groundwater filter velocity vector part in y direction (m/s). Path for new raster layer.

GRASS_REGION_PARAMETER

extent - GRASS GIS region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..

GRASS_REGION_CELLSIZE_PARAMETER

number - GRASS GIS region cellsize (leave 0 for default). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

GRASS_RASTER_FORMAT_OPT

string - Output Rasters format options (createopt). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

GRASS_RASTER_FORMAT_META

string - Output Rasters format metadata options (metaopt). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

...

further parameters passed to qgisprocess::qgis_run_algorithm()

.complete_output

logical specifying if complete out of qgisprocess::qgis_run_algorithm() should be used (TRUE) or first output (most likely the main) should read (FALSE). Default value is TRUE.

.quiet

logical specifying if parameter .quiet for qgisprocess::qgis_run_algorithm() Default value is TRUE.

.messages

logical specifying if messages from qgisprocess::qgis_run_algorithm() should be printed (TRUE) or not (FALSE). Default value is FALSE.

Details

Outputs description

output - outputRaster - Solute Transport

vx - outputRaster - Calculate and store the groundwater filter velocity vector part in x direction (m/s)

vy - outputRaster - Calculate and store the groundwater filter velocity vector part in y direction (m/s)