QGIS algorithm - Coordinate transformation (grid)

QGIS Algorithm provided by SAGA Next Gen Coordinate transformation (grid) (sagang:coordinatetransformationgrid). —————- Arguments —————- CRS_METHOD: Get CRS Definition from... Default value: 0 Argument type: enum Available values: - 0: (0) Proj4 Parameters - 1: (1) EPSG Code - 2: (2) Well Known Text File Acceptable values: - Number of selected option, e.g. '1' - Comma separated list of options, e.g. '1,3' CRS_PROJ4: Proj4 Parameters Default value: Argument type: string Acceptable values: - String value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression CRS_FILE: Well Known Text File Default value: False Argument type: file Acceptable values: - Path to a file CRS_EPSG: Authority Code Default value: -1 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression CRS_EPSG_AUTH: Authority Default value: EPSG Argument type: string Acceptable values: - String value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SOURCE: Source Argument type: raster Acceptable values: - Path to a raster layer RESAMPLING: Resampling Default value: 3 Argument type: enum Available values: - 0: (0) Nearest Neighbour - 1: (1) Bilinear Interpolation - 2: (2) Bicubic Spline Interpolation - 3: (3) B-Spline Interpolation Acceptable values: - Number of selected option, e.g. '1' - Comma separated list of options, e.g. '1,3' BYTEWISE: Bytewise Interpolation Default value: false Argument type: boolean Acceptable values: - 1 for true/yes - 0 for false/no - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression KEEP_TYPE: Preserve Data Type Default value: false Argument type: boolean Acceptable values: - 1 for true/yes - 0 for false/no - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression TARGET_AREA: Use Target Area Polygon Default value: false Argument type: boolean Acceptable values: - 1 for true/yes - 0 for false/no - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression TARGET_USER_XMIN TARGET_USER_XMAX TARGET_USER_YMIN TARGET_USER_YMAX: Output extent (optional) Argument type: extent Acceptable values: - A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105' - Path to a layer. The extent of the layer is used. TARGET_USER_SIZE: Cellsize Default value: 100 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression GRID: Target Argument type: rasterDestination Acceptable values: - Path for new raster layer OUT_X: X Coordinates Argument type: rasterDestination Acceptable values: - Path for new raster layer OUT_Y: Y Coordinates Argument type: rasterDestination Acceptable values: - Path for new raster layer

sagang_coordinatetransformationgrid(
  CRS_METHOD = qgisprocess:::qgis_default_value(),
  CRS_PROJ4 = qgisprocess:::qgis_default_value(),
  CRS_FILE = qgisprocess:::qgis_default_value(),
  CRS_EPSG = qgisprocess:::qgis_default_value(),
  CRS_EPSG_AUTH = qgisprocess:::qgis_default_value(),
  SOURCE = qgisprocess:::qgis_default_value(),
  RESAMPLING = qgisprocess:::qgis_default_value(),
  BYTEWISE = qgisprocess:::qgis_default_value(),
  KEEP_TYPE = qgisprocess:::qgis_default_value(),
  TARGET_AREA = qgisprocess:::qgis_default_value(),
  TARGET_USER_XMIN_TARGET_USER_XMAX_TARGET_USER_YMIN_TARGET_USER_YMAX =
    qgisprocess:::qgis_default_value(),
  TARGET_USER_SIZE = qgisprocess:::qgis_default_value(),
  GRID = qgisprocess:::qgis_default_value(),
  OUT_X = qgisprocess:::qgis_default_value(),
  OUT_Y = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)

Arguments

CRS_METHOD

enum of ("(0) Proj4 Parameters", "(1) EPSG Code", "(2) Well Known Text File") - Get CRS Definition from.... Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.

CRS_PROJ4

string - Proj4 Parameters. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

CRS_FILE

file - Well Known Text File. Path to a file.

CRS_EPSG

number - Authority Code. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

CRS_EPSG_AUTH

string - Authority. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

SOURCE

raster - Source. Path to a raster layer.

RESAMPLING

enum of ("(0) Nearest Neighbour", "(1) Bilinear Interpolation", "(2) Bicubic Spline Interpolation", "(3) B-Spline Interpolation") - Resampling. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.

BYTEWISE

boolean - Bytewise Interpolation. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

KEEP_TYPE

boolean - Preserve Data Type. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

TARGET_AREA

boolean - Use Target Area Polygon. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

TARGET_USER_XMIN_TARGET_USER_XMAX_TARGET_USER_YMIN_TARGET_USER_YMAX

extent - Output extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used.. Original algorithm parameter name: TARGET_USER_XMIN TARGET_USER_XMAX TARGET_USER_YMIN TARGET_USER_YMAX.

TARGET_USER_SIZE

number - Cellsize. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

GRID

rasterDestination - Target. Path for new raster layer.

OUT_X

rasterDestination - X Coordinates. Path for new raster layer.

OUT_Y

rasterDestination - Y Coordinates. Path for new raster layer.

...

further parameters passed to qgisprocess::qgis_run_algorithm()

.complete_output

logical specifying if complete out of qgisprocess::qgis_run_algorithm() should be used (TRUE) or first output (most likely the main) should read (FALSE). Default value is TRUE.

.quiet

logical specifying if parameter .quiet for qgisprocess::qgis_run_algorithm() Default value is TRUE.

.messages

logical specifying if messages from qgisprocess::qgis_run_algorithm() should be printed (TRUE) or not (FALSE). Default value is FALSE.

Details

Outputs description

GRID - outputRaster - Target

OUT_X - outputRaster - X Coordinates

OUT_Y - outputRaster - Y Coordinates