QGIS algorithm - Gwr for multiple predictor layers

QGIS Algorithm provided by SAGA Next Gen Gwr for multiple predictor layers (sagang:gwrformultiplepredictorlayers). —————- Arguments —————- POINTS: Points Argument type: source Acceptable values: - Path to a vector layer DEPENDENT: Dependent Variable Argument type: field Acceptable values: - The name of an existing field - ; delimited list of existing field names RESIDUALS: Residuals Argument type: vectorDestination Acceptable values: - Path for new vector layer PREDICTORS: Predictors Argument type: multilayer REGRESSION: Regression Argument type: rasterDestination Acceptable values: - Path for new raster layer QUALITY: Coefficient of Determination Argument type: rasterDestination Acceptable values: - Path for new raster layer MODEL: Regression Parameters Argument type: rasterDestination Acceptable values: - Path for new raster layer LOGISTIC: Logistic Regression Default value: false Argument type: boolean Acceptable values: - 1 for true/yes - 0 for false/no - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression MODEL_OUT: Output of Regression Parameters Default value: false Argument type: boolean Acceptable values: - 1 for true/yes - 0 for false/no - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression RESOLUTION: Model Resolution Default value: 1 Argument type: enum Available values: - 0: (0) same as predictors - 1: (1) user defined Acceptable values: - Number of selected option, e.g. '1' - Comma separated list of options, e.g. '1,3' RESOLUTION_VAL: Resolution Default value: 1 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DW_WEIGHTING: Weighting Function Default value: 3 Argument type: enum Available values: - 0: (0) no distance weighting - 1: (1) inverse distance to a power - 2: (2) exponential - 3: (3) gaussian Acceptable values: - Number of selected option, e.g. '1' - Comma separated list of options, e.g. '1,3' DW_IDW_POWER: Power Default value: 2 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DW_BANDWIDTH: Bandwidth Default value: 1 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SEARCH_RANGE: Search Range Default value: 1 Argument type: enum Available values: - 0: (0) local - 1: (1) global Acceptable values: - Number of selected option, e.g. '1' - Comma separated list of options, e.g. '1,3' SEARCH_RADIUS: Maximum Search Distance Default value: 1000 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SEARCH_POINTS_ALL: Number of Points Default value: 1 Argument type: enum Available values: - 0: (0) maximum number of nearest points - 1: (1) all points within search distance Acceptable values: - Number of selected option, e.g. '1' - Comma separated list of options, e.g. '1,3' SEARCH_POINTS_MIN: Minimum Default value: 16 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SEARCH_POINTS_MAX: Maximum Default value: 20 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SEARCH_DIRECTION: Direction Default value: 0 Argument type: enum Available values: - 0: (0) all directions - 1: (1) quadrants Acceptable values: - Number of selected option, e.g. '1' - Comma separated list of options, e.g. '1,3'

sagang_gwrformultiplepredictorlayers(
  POINTS = qgisprocess:::qgis_default_value(),
  DEPENDENT = qgisprocess:::qgis_default_value(),
  RESIDUALS = qgisprocess:::qgis_default_value(),
  PREDICTORS = qgisprocess:::qgis_default_value(),
  REGRESSION = qgisprocess:::qgis_default_value(),
  QUALITY = qgisprocess:::qgis_default_value(),
  MODEL = qgisprocess:::qgis_default_value(),
  LOGISTIC = qgisprocess:::qgis_default_value(),
  MODEL_OUT = qgisprocess:::qgis_default_value(),
  RESOLUTION = qgisprocess:::qgis_default_value(),
  RESOLUTION_VAL = qgisprocess:::qgis_default_value(),
  DW_WEIGHTING = qgisprocess:::qgis_default_value(),
  DW_IDW_POWER = qgisprocess:::qgis_default_value(),
  DW_BANDWIDTH = qgisprocess:::qgis_default_value(),
  SEARCH_RANGE = qgisprocess:::qgis_default_value(),
  SEARCH_RADIUS = qgisprocess:::qgis_default_value(),
  SEARCH_POINTS_ALL = qgisprocess:::qgis_default_value(),
  SEARCH_POINTS_MIN = qgisprocess:::qgis_default_value(),
  SEARCH_POINTS_MAX = qgisprocess:::qgis_default_value(),
  SEARCH_DIRECTION = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)

Arguments

POINTS

source - Points. Path to a vector layer.

DEPENDENT

field - Dependent Variable. The name of an existing field. ; delimited list of existing field names.

RESIDUALS

vectorDestination - Residuals. Path for new vector layer.

PREDICTORS

multilayer - Predictors. .

REGRESSION

rasterDestination - Regression. Path for new raster layer.

QUALITY

rasterDestination - Coefficient of Determination. Path for new raster layer.

MODEL

rasterDestination - Regression Parameters. Path for new raster layer.

LOGISTIC

boolean - Logistic Regression. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

MODEL_OUT

boolean - Output of Regression Parameters. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

RESOLUTION

enum of ("(0) same as predictors", "(1) user defined") - Model Resolution. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.

RESOLUTION_VAL

number - Resolution. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

DW_WEIGHTING

enum of ("(0) no distance weighting", "(1) inverse distance to a power", "(2) exponential", "(3) gaussian") - Weighting Function. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.

DW_IDW_POWER

number - Power. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

DW_BANDWIDTH

number - Bandwidth. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

SEARCH_RANGE

enum of ("(0) local", "(1) global") - Search Range. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.

SEARCH_RADIUS

number - Maximum Search Distance. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

SEARCH_POINTS_ALL

enum of ("(0) maximum number of nearest points", "(1) all points within search distance") - Number of Points. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.

SEARCH_POINTS_MIN

number - Minimum. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

SEARCH_POINTS_MAX

number - Maximum. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

SEARCH_DIRECTION

enum of ("(0) all directions", "(1) quadrants") - Direction. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.

...

further parameters passed to qgisprocess::qgis_run_algorithm()

.complete_output

logical specifying if complete out of qgisprocess::qgis_run_algorithm() should be used (TRUE) or first output (most likely the main) should read (FALSE). Default value is TRUE.

.quiet

logical specifying if parameter .quiet for qgisprocess::qgis_run_algorithm() Default value is TRUE.

.messages

logical specifying if messages from qgisprocess::qgis_run_algorithm() should be printed (TRUE) or not (FALSE). Default value is FALSE.

Details

Outputs description

MODEL - outputRaster - Regression Parameters

QUALITY - outputRaster - Coefficient of Determination

REGRESSION - outputRaster - Regression

RESIDUALS - outputVector - Residuals