QGIS algorithm - Landflow version 1.0 (build 3.5.1b)

QGIS Algorithm provided by SAGA Next Gen Landflow version 1.0 (build 3.5.1b) (sagang:landflowversion10build351b). —————- Arguments —————- INPUT: Gelaendemodell (DTM) Argument type: raster Acceptable values: - Path to a raster layer INPUT2: Hoehengradienten (Grad) Argument type: raster Acceptable values: - Path to a raster layer INPUT10: Flusseinzugsgebiet (BasinShare) Argument type: raster Acceptable values: - Path to a raster layer INPUT8: HG-Raster (HGGrid) Argument type: raster Acceptable values: - Path to a raster layer INPUT9: HG-Hoehengradienten (HGGrad) Argument type: raster Acceptable values: - Path to a raster layer INPUT11: Referenzverlauf Flussbilanz (optional) Argument type: raster Acceptable values: - Path to a raster layer INPUT12: Zuflussrasterzellen (NumInFlowCells) (optional) Argument type: raster Acceptable values: - Path to a raster layer INPUT3: statische Wassserentnahme (statWUse) (optional) Argument type: raster Acceptable values: - Path to a raster layer INPUT5: Fluss-Speicher einlesen (optional) Argument type: raster Acceptable values: - Path to a raster layer INPUT6: Oberflaechenabfluss-Speicher einlesen (optional) Argument type: raster Acceptable values: - Path to a raster layer INPUT7: Grundwasserabfluss-Speicher einlesen (optional) Argument type: raster Acceptable values: - Path to a raster layer OUTPUT: Fluss-Speicher ausgeben Argument type: rasterDestination Acceptable values: - Path for new raster layer OUTPUT2: Oberflaechenabfluss-Speicher ausgeben Argument type: rasterDestination Acceptable values: - Path for new raster layer OUTPUT3: Grundwasserabfluss-Speicher ausgeben Argument type: rasterDestination Acceptable values: - Path for new raster layer OUTPUT4: Wasserflussvolumen Argument type: rasterDestination Acceptable values: - Path for new raster layer OUTPUT5: SumRunoffDrainage Argument type: rasterDestination Acceptable values: - Path for new raster layer OUTPUT6: DynWUse Argument type: rasterDestination Acceptable values: - Path for new raster layer onlyRB: Berechnung nur im Flusseinzugsgebiet Default value: true Argument type: boolean Acceptable values: - 1 for true/yes - 0 for false/no - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression TimeStep: Zeitschrittdauer (s) Default value: 1 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression CalcT: automatisierter max. Zeitschritt Default value: true Argument type: boolean Acceptable values: - 1 for true/yes - 0 for false/no - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression sYear: Startjahr Default value: 1999 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DayNum: Anzahl der Simulationstage Default value: 365 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression Folder2: Speicherordner Default value: False Argument type: file Acceptable values: - Path to a file Folder1: Pfad LS-Daten Default value: False Argument type: file Acceptable values: - Path to a file stRedFacR: Reduzierungsfaktor Surface Runoff (% / 100) Default value: 0 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression autoFacD: Ausgleich der Surface Runoff Reduktion Default value: false Argument type: boolean Acceptable values: - 1 for true/yes - 0 for false/no - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression stRedFacD: Reduzierungsfaktor Drainage (% / 100) Default value: 0 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression OffsetR: Offsetwert Surface Runoff (m3/s) Default value: 0 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression OffsetD: Offsetwert Drainage (m3/s) Default value: 0 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression CacheUse: Initiierung mit vorhandenen Speichern Default value: false Argument type: boolean Acceptable values: - 1 for true/yes - 0 for false/no - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression Folder4: NCache einlesen Default value: False Argument type: file Acceptable values: - Path to a file wNC: Ausgabe NCache Default value: false Argument type: boolean Acceptable values: - 1 for true/yes - 0 for false/no - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression ParamC: Gerinne-Parameter cG Default value: 0.06 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression ParamG: Oberflaechen-Parameter cO Default value: 0.357 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression ParamB: Grundwasser-Parameter pB Default value: 300 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression nG: Gerinne-Speicherkaskade nG Default value: 3 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression RivG: Beruecksichtigung der Hauptgerinnerasterzellen? Default value: 1 Argument type: enum Available values: - 0: (0) nein - 1: (1) ja - 2: bestimmen anhand eines Hauptgerinnerasters (HG Raster) Acceptable values: - Number of selected option, e.g. '1' - Comma separated list of options, e.g. '1,3' ParamCr: Hauptgerinne-Parameter cHG Default value: 0.007 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression nHG: Hauptgerinne-Speicherkaskade nHG Default value: 1 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression EnfVmax: Abflussgeschwindigkeit begrenzen Default value: true Argument type: boolean Acceptable values: - 1 for true/yes - 0 for false/no - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression VTresh: Oberer Abflussgeschwindigkeitsgrenzwert (km/h) Default value: 4 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression WCons: Dynamische Flaechenwassernutzung... Default value: 0 Argument type: enum Available values: - 0: (0) keine - 1: (1) ...anteilig aus den Hauptgerinnerasterzellen - 2: (2) ...anteilig aus Rasterzellen der Teileinzugegebiete - 3: (3) ...genau den entsprechenden Wert aus der jeweiligen Rasterzelle entnehmen Acceptable values: - Number of selected option, e.g. '1' - Comma separated list of options, e.g. '1,3' Folder3: Ordnerpfad der WUse Daten Default value: False Argument type: file Acceptable values: - Path to a file WConUnit: WUse Einheit Default value: 0 Argument type: enum Available values: - 0: (0) m3/s - 1: (1) m3/Monat Acceptable values: - Number of selected option, e.g. '1' - Comma separated list of options, e.g. '1,3' WConsD: Vorlauftage ohne Entnahmen Default value: 0 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression WConThres: Abflussschwellenwert (m3/s) Default value: 0 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression stConsAll: Abflussreduzierungsfaktor Gerinnerasterzellen (% / 100) Default value: 0 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression stConsRiv: Abflussreduzierungsfaktor HG-Rasterzellen (% / 100) Default value: 0 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression vRM1: Abflussmanipulation Rasterzelle 1... Default value: 0 Argument type: enum Available values: - 0: (0) nein - 1: (1) Ja, berechneten Abfluss veraendern: res. Abfluss = berechn. Abfluss * q + a - 2: (2) Ja, Abfluss manuell vorgeben: res. Abfluss = Speicherinhalt * q + a Acceptable values: - Number of selected option, e.g. '1' - Comma separated list of options, e.g. '1,3' RM1x: Rasterzelle 1 (x-Koord.) Default value: -1 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression RM1y: Rasterzelle 1 (y-Koord.) Default value: -1 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression RM1q: q (%/100) Default value: 1 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression RM1a: a (+-m3/s) Default value: 0 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression vRM2: Abflussmanipulation Rasterzelle 2... Default value: 0 Argument type: enum Available values: - 0: (0) nein - 1: (1) Ja, berechneten Abfluss veraendern: res. Abfluss = berechn. Abfluss * q + a - 2: (2) Ja, Abfluss manuell vorgeben: res. Abfluss = Speicherinhalt * q + a Acceptable values: - Number of selected option, e.g. '1' - Comma separated list of options, e.g. '1,3' RM2x: Rasterzelle 2 (x-Koord.) Default value: -1 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression RM2y: Rasterzelle 2 (y-Koord.) Default value: -1 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression RM2q: q (%/100) Default value: 1 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression RM2a: a (+-m3/s) Default value: 0 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression EvP1s: EvP1 Name Default value: NeuDarchau.txt Argument type: string Acceptable values: - String value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression EvP1x: EvP1 Rasterzelle (x-Koord.) Default value: 30 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression EvP1y: EvP1 Rasterzelle (y-Koord.) Default value: 115 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression EvP2s: EvP2 Name Default value: Lutherstadt-Wittenberg.txt Argument type: string Acceptable values: - String value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression EvP2x: EvP2 Rasterzelle (x-Koord.) Default value: 54 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression EvP2y: EvP2 Rasterzelle (y-Koord.) Default value: 85 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression EvP3s: EvP3 Name Default value: Schoena.txt Argument type: string Acceptable values: - String value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression EvP3x: EvP3 Rasterzelle (x-Koord.) Default value: 78 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression EvP3y: EvP3 Rasterzelle (y-Koord.) Default value: 65 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression RBMx: Ausgangsrasterzelle (x-Koord.) Default value: 16 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression RBMy: Ausgangsrasterzelle (y-Koord.) Default value: 121 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression wP: SimParameters Default value: true Argument type: boolean Acceptable values: - 1 for true/yes - 0 for false/no - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression eP: Fehlerprotokoll Default value: true Argument type: boolean Acceptable values: - 1 for true/yes - 0 for false/no - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression MoniLog1: RiverBasinDay-Monitoring Default value: true Argument type: boolean Acceptable values: - 1 for true/yes - 0 for false/no - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression MoniLog3: RiverBasinMonth-Monitoring Default value: true Argument type: boolean Acceptable values: - 1 for true/yes - 0 for false/no - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression MoniLog2: WSystem-Monitoring Default value: true Argument type: boolean Acceptable values: - 1 for true/yes - 0 for false/no - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression Test1: Testroutine1 durchfuehren... Default value: 0 Argument type: enum Available values: - 0: (0) nein - 1: (1) Ja, TestRoutine1 nur fuer Teileinzugsgbiet der HG-Rasterzelle - 2: (2) Ja, TestRoutine1 fuer Flusseinzugsgebiet bis zu der HG-Rasterzelle Acceptable values: - Number of selected option, e.g. '1' - Comma separated list of options, e.g. '1,3' xt1: Hauptgerinnerasterzelle (x-Koord.) Default value: 0 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression yt1: Hauptgerinnerasterzelle (y-Koord.) Default value: 0 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression

sagang_landflowversion10build351b(
  INPUT = qgisprocess:::qgis_default_value(),
  INPUT2 = qgisprocess:::qgis_default_value(),
  INPUT10 = qgisprocess:::qgis_default_value(),
  INPUT8 = qgisprocess:::qgis_default_value(),
  INPUT9 = qgisprocess:::qgis_default_value(),
  INPUT11 = qgisprocess:::qgis_default_value(),
  INPUT12 = qgisprocess:::qgis_default_value(),
  INPUT3 = qgisprocess:::qgis_default_value(),
  INPUT5 = qgisprocess:::qgis_default_value(),
  INPUT6 = qgisprocess:::qgis_default_value(),
  INPUT7 = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  OUTPUT2 = qgisprocess:::qgis_default_value(),
  OUTPUT3 = qgisprocess:::qgis_default_value(),
  OUTPUT4 = qgisprocess:::qgis_default_value(),
  OUTPUT5 = qgisprocess:::qgis_default_value(),
  OUTPUT6 = qgisprocess:::qgis_default_value(),
  onlyRB = qgisprocess:::qgis_default_value(),
  TimeStep = qgisprocess:::qgis_default_value(),
  CalcT = qgisprocess:::qgis_default_value(),
  sYear = qgisprocess:::qgis_default_value(),
  DayNum = qgisprocess:::qgis_default_value(),
  Folder2 = qgisprocess:::qgis_default_value(),
  Folder1 = qgisprocess:::qgis_default_value(),
  stRedFacR = qgisprocess:::qgis_default_value(),
  autoFacD = qgisprocess:::qgis_default_value(),
  stRedFacD = qgisprocess:::qgis_default_value(),
  OffsetR = qgisprocess:::qgis_default_value(),
  OffsetD = qgisprocess:::qgis_default_value(),
  CacheUse = qgisprocess:::qgis_default_value(),
  Folder4 = qgisprocess:::qgis_default_value(),
  wNC = qgisprocess:::qgis_default_value(),
  ParamC = qgisprocess:::qgis_default_value(),
  ParamG = qgisprocess:::qgis_default_value(),
  ParamB = qgisprocess:::qgis_default_value(),
  nG = qgisprocess:::qgis_default_value(),
  RivG = qgisprocess:::qgis_default_value(),
  ParamCr = qgisprocess:::qgis_default_value(),
  nHG = qgisprocess:::qgis_default_value(),
  EnfVmax = qgisprocess:::qgis_default_value(),
  VTresh = qgisprocess:::qgis_default_value(),
  WCons = qgisprocess:::qgis_default_value(),
  Folder3 = qgisprocess:::qgis_default_value(),
  WConUnit = qgisprocess:::qgis_default_value(),
  WConsD = qgisprocess:::qgis_default_value(),
  WConThres = qgisprocess:::qgis_default_value(),
  stConsAll = qgisprocess:::qgis_default_value(),
  stConsRiv = qgisprocess:::qgis_default_value(),
  vRM1 = qgisprocess:::qgis_default_value(),
  RM1x = qgisprocess:::qgis_default_value(),
  RM1y = qgisprocess:::qgis_default_value(),
  RM1q = qgisprocess:::qgis_default_value(),
  RM1a = qgisprocess:::qgis_default_value(),
  vRM2 = qgisprocess:::qgis_default_value(),
  RM2x = qgisprocess:::qgis_default_value(),
  RM2y = qgisprocess:::qgis_default_value(),
  RM2q = qgisprocess:::qgis_default_value(),
  RM2a = qgisprocess:::qgis_default_value(),
  EvP1s = qgisprocess:::qgis_default_value(),
  EvP1x = qgisprocess:::qgis_default_value(),
  EvP1y = qgisprocess:::qgis_default_value(),
  EvP2s = qgisprocess:::qgis_default_value(),
  EvP2x = qgisprocess:::qgis_default_value(),
  EvP2y = qgisprocess:::qgis_default_value(),
  EvP3s = qgisprocess:::qgis_default_value(),
  EvP3x = qgisprocess:::qgis_default_value(),
  EvP3y = qgisprocess:::qgis_default_value(),
  RBMx = qgisprocess:::qgis_default_value(),
  RBMy = qgisprocess:::qgis_default_value(),
  wP = qgisprocess:::qgis_default_value(),
  eP = qgisprocess:::qgis_default_value(),
  MoniLog1 = qgisprocess:::qgis_default_value(),
  MoniLog3 = qgisprocess:::qgis_default_value(),
  MoniLog2 = qgisprocess:::qgis_default_value(),
  Test1 = qgisprocess:::qgis_default_value(),
  xt1 = qgisprocess:::qgis_default_value(),
  yt1 = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)

Arguments

INPUT

raster - Gelaendemodell (DTM). Path to a raster layer.

INPUT2

raster - Hoehengradienten (Grad). Path to a raster layer.

INPUT10

raster - Flusseinzugsgebiet (BasinShare). Path to a raster layer.

INPUT8

raster - HG-Raster (HGGrid). Path to a raster layer.

INPUT9

raster - HG-Hoehengradienten (HGGrad). Path to a raster layer.

INPUT11

raster - Referenzverlauf Flussbilanz. Path to a raster layer.

INPUT12

raster - Zuflussrasterzellen (NumInFlowCells). Path to a raster layer.

INPUT3

raster - statische Wassserentnahme (statWUse). Path to a raster layer.

INPUT5

raster - Fluss-Speicher einlesen. Path to a raster layer.

INPUT6

raster - Oberflaechenabfluss-Speicher einlesen. Path to a raster layer.

INPUT7

raster - Grundwasserabfluss-Speicher einlesen. Path to a raster layer.

OUTPUT

rasterDestination - Fluss-Speicher ausgeben. Path for new raster layer.

OUTPUT2

rasterDestination - Oberflaechenabfluss-Speicher ausgeben. Path for new raster layer.

OUTPUT3

rasterDestination - Grundwasserabfluss-Speicher ausgeben. Path for new raster layer.

OUTPUT4

rasterDestination - Wasserflussvolumen. Path for new raster layer.

OUTPUT5

rasterDestination - SumRunoffDrainage. Path for new raster layer.

OUTPUT6

rasterDestination - DynWUse. Path for new raster layer.

onlyRB

boolean - Berechnung nur im Flusseinzugsgebiet. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

TimeStep

number - Zeitschrittdauer (s). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

CalcT

boolean - automatisierter max. Zeitschritt. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

sYear

number - Startjahr. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

DayNum

number - Anzahl der Simulationstage. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

Folder2

file - Speicherordner. Path to a file.

Folder1

file - Pfad LS-Daten. Path to a file.

stRedFacR

number - Reduzierungsfaktor Surface Runoff (% / 100). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

autoFacD

boolean - Ausgleich der Surface Runoff Reduktion. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

stRedFacD

number - Reduzierungsfaktor Drainage (% / 100). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

OffsetR

number - Offsetwert Surface Runoff (m3/s). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

OffsetD

number - Offsetwert Drainage (m3/s). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

CacheUse

boolean - Initiierung mit vorhandenen Speichern. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

Folder4

file - NCache einlesen. Path to a file.

wNC

boolean - Ausgabe NCache. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

ParamC

number - Gerinne-Parameter cG. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

ParamG

number - Oberflaechen-Parameter cO. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

ParamB

number - Grundwasser-Parameter pB. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

nG

number - Gerinne-Speicherkaskade nG. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

RivG

enum of ("(0) nein", "(1) ja", " bestimmen anhand eines Hauptgerinnerasters (HG Raster)") - Beruecksichtigung der Hauptgerinnerasterzellen?. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.

ParamCr

number - Hauptgerinne-Parameter cHG. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

nHG

number - Hauptgerinne-Speicherkaskade nHG. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

EnfVmax

boolean - Abflussgeschwindigkeit begrenzen. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

VTresh

number - Oberer Abflussgeschwindigkeitsgrenzwert (km/h). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

WCons

enum of ("(0) keine", "(1) ...anteilig aus den Hauptgerinnerasterzellen", "(2) ...anteilig aus Rasterzellen der Teileinzugegebiete", "(3) ...genau den entsprechenden Wert aus der jeweiligen Rasterzelle entnehmen") - Dynamische Flaechenwassernutzung.... Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.

Folder3

file - Ordnerpfad der WUse Daten. Path to a file.

WConUnit

enum of ("(0) m3/s", "(1) m3/Monat") - WUse Einheit. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.

WConsD

number - Vorlauftage ohne Entnahmen. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

WConThres

number - Abflussschwellenwert (m3/s). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

stConsAll

number - Abflussreduzierungsfaktor Gerinnerasterzellen (% / 100). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

stConsRiv

number - Abflussreduzierungsfaktor HG-Rasterzellen (% / 100). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

vRM1

enum of ("(0) nein", "(1) Ja, berechneten Abfluss veraendern: res. Abfluss = berechn. Abfluss * q + a", "(2) Ja, Abfluss manuell vorgeben: res. Abfluss = Speicherinhalt * q + a") - Abflussmanipulation Rasterzelle 1... . Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.

RM1x

number - Rasterzelle 1 (x-Koord.). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

RM1y

number - Rasterzelle 1 (y-Koord.). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

RM1q

number - q (%/100). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

RM1a

number - a (+-m3/s). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

vRM2

enum of ("(0) nein", "(1) Ja, berechneten Abfluss veraendern: res. Abfluss = berechn. Abfluss * q + a", "(2) Ja, Abfluss manuell vorgeben: res. Abfluss = Speicherinhalt * q + a") - Abflussmanipulation Rasterzelle 2... . Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.

RM2x

number - Rasterzelle 2 (x-Koord.). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

RM2y

number - Rasterzelle 2 (y-Koord.). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

RM2q

number - q (%/100). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

RM2a

number - a (+-m3/s). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

EvP1s

string - EvP1 Name. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

EvP1x

number - EvP1 Rasterzelle (x-Koord.). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

EvP1y

number - EvP1 Rasterzelle (y-Koord.). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

EvP2s

string - EvP2 Name. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

EvP2x

number - EvP2 Rasterzelle (x-Koord.). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

EvP2y

number - EvP2 Rasterzelle (y-Koord.). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

EvP3s

string - EvP3 Name. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

EvP3x

number - EvP3 Rasterzelle (x-Koord.). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

EvP3y

number - EvP3 Rasterzelle (y-Koord.). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

RBMx

number - Ausgangsrasterzelle (x-Koord.). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

RBMy

number - Ausgangsrasterzelle (y-Koord.). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

wP

boolean - SimParameters. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

eP

boolean - Fehlerprotokoll. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

MoniLog1

boolean - RiverBasinDay-Monitoring. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

MoniLog3

boolean - RiverBasinMonth-Monitoring. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

MoniLog2

boolean - WSystem-Monitoring. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

Test1

enum of ("(0) nein", "(1) Ja, TestRoutine1 nur fuer Teileinzugsgbiet der HG-Rasterzelle", "(2) Ja, TestRoutine1 fuer Flusseinzugsgebiet bis zu der HG-Rasterzelle") - Testroutine1 durchfuehren... . Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.

xt1

number - Hauptgerinnerasterzelle (x-Koord.). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

yt1

number - Hauptgerinnerasterzelle (y-Koord.). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

...

further parameters passed to qgisprocess::qgis_run_algorithm()

.complete_output

logical specifying if complete out of qgisprocess::qgis_run_algorithm() should be used (TRUE) or first output (most likely the main) should read (FALSE). Default value is TRUE.

.quiet

logical specifying if parameter .quiet for qgisprocess::qgis_run_algorithm() Default value is TRUE.

.messages

logical specifying if messages from qgisprocess::qgis_run_algorithm() should be printed (TRUE) or not (FALSE). Default value is FALSE.

Details

Outputs description

OUTPUT - outputRaster - Fluss-Speicher ausgeben

OUTPUT2 - outputRaster - Oberflaechenabfluss-Speicher ausgeben

OUTPUT3 - outputRaster - Grundwasserabfluss-Speicher ausgeben

OUTPUT4 - outputRaster - Wasserflussvolumen

OUTPUT5 - outputRaster - SumRunoffDrainage

OUTPUT6 - outputRaster - DynWUse