QGIS algorithm - Thin plate spline

QGIS Algorithm provided by SAGA Next Gen Thin plate spline (sagang:thinplatespline). —————- Arguments —————- SHAPES: Points Argument type: source Acceptable values: - Path to a vector layer FIELD: Attribute Argument type: field Acceptable values: - The name of an existing field - ; delimited list of existing field names TARGET_USER_XMIN TARGET_USER_XMAX TARGET_USER_YMIN TARGET_USER_YMAX: Output extent (optional) Argument type: extent Acceptable values: - A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105' - Path to a layer. The extent of the layer is used. TARGET_USER_SIZE: Cellsize Default value: 100 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression TARGET_OUT_GRID: Target Grid Argument type: rasterDestination Acceptable values: - Path for new raster layer REGULARISATION: Regularisation Default value: 0.0001 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SEARCH_RANGE: Search Range Default value: 1 Argument type: enum Available values: - 0: (0) local - 1: (1) global Acceptable values: - Number of selected option, e.g. '1' - Comma separated list of options, e.g. '1,3' SEARCH_RADIUS: Maximum Search Distance Default value: 1000 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SEARCH_POINTS_ALL: Number of Points Default value: 1 Argument type: enum Available values: - 0: (0) maximum number of nearest points - 1: (1) all points within search distance Acceptable values: - Number of selected option, e.g. '1' - Comma separated list of options, e.g. '1,3' SEARCH_POINTS_MIN: Minimum Default value: 16 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SEARCH_POINTS_MAX: Maximum Default value: 20 Argument type: number Acceptable values: - A numeric value - field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field - expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SEARCH_DIRECTION: Direction Default value: 0 Argument type: enum Available values: - 0: (0) all directions - 1: (1) quadrants Acceptable values: - Number of selected option, e.g. '1' - Comma separated list of options, e.g. '1,3'

sagang_thinplatespline(
  SHAPES = qgisprocess:::qgis_default_value(),
  FIELD = qgisprocess:::qgis_default_value(),
  TARGET_USER_XMIN_TARGET_USER_XMAX_TARGET_USER_YMIN_TARGET_USER_YMAX =
    qgisprocess:::qgis_default_value(),
  TARGET_USER_SIZE = qgisprocess:::qgis_default_value(),
  TARGET_OUT_GRID = qgisprocess:::qgis_default_value(),
  REGULARISATION = qgisprocess:::qgis_default_value(),
  SEARCH_RANGE = qgisprocess:::qgis_default_value(),
  SEARCH_RADIUS = qgisprocess:::qgis_default_value(),
  SEARCH_POINTS_ALL = qgisprocess:::qgis_default_value(),
  SEARCH_POINTS_MIN = qgisprocess:::qgis_default_value(),
  SEARCH_POINTS_MAX = qgisprocess:::qgis_default_value(),
  SEARCH_DIRECTION = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)

Arguments

SHAPES

source - Points. Path to a vector layer.

FIELD

field - Attribute. The name of an existing field. ; delimited list of existing field names.

TARGET_USER_XMIN_TARGET_USER_XMAX_TARGET_USER_YMIN_TARGET_USER_YMAX

extent - Output extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used.. Original algorithm parameter name: TARGET_USER_XMIN TARGET_USER_XMAX TARGET_USER_YMIN TARGET_USER_YMAX.

TARGET_USER_SIZE

number - Cellsize. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

TARGET_OUT_GRID

rasterDestination - Target Grid. Path for new raster layer.

REGULARISATION

number - Regularisation. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

SEARCH_RANGE

enum of ("(0) local", "(1) global") - Search Range. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.

SEARCH_RADIUS

number - Maximum Search Distance. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

SEARCH_POINTS_ALL

enum of ("(0) maximum number of nearest points", "(1) all points within search distance") - Number of Points. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.

SEARCH_POINTS_MIN

number - Minimum. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

SEARCH_POINTS_MAX

number - Maximum. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.

SEARCH_DIRECTION

enum of ("(0) all directions", "(1) quadrants") - Direction. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.

...

further parameters passed to qgisprocess::qgis_run_algorithm()

.complete_output

logical specifying if complete out of qgisprocess::qgis_run_algorithm() should be used (TRUE) or first output (most likely the main) should read (FALSE). Default value is TRUE.

.quiet

logical specifying if parameter .quiet for qgisprocess::qgis_run_algorithm() Default value is TRUE.

.messages

logical specifying if messages from qgisprocess::qgis_run_algorithm() should be printed (TRUE) or not (FALSE). Default value is FALSE.

Details

Outputs description

TARGET_OUT_GRID - outputRaster - Target Grid